backend/requirements.txt
twitchio==2.6.0
pymongo==4.6.1
python-dotenv==1.0.0
aiohttp==3.9.1 